requirements.txt
google-generativeai
streamlit==1.29.0
GoogleIt